Add register dump function and instruction execution loop
A function to print the current values of all registers (dump_registers) has been added. Furthermore, the main function now contains a loop to continuously fetch and execute instructions from memory. This will continue until a non-implemented instruction is encountered or the end of program marker is reached, at which point the emulator is destroyed and the program ends.