Add more emulator instructions and create ModRM struct
The commit includes the addition of several new functions to the emulator, which implement various instructions. Additionally, a new struct, ModRM, has been created to parse modrm bytes from the instruction stream. The added functions handle movement and arithmetic operations, as well as the new struct for memory addressing. This constitutes a significant expansion of the emulator's capacities that equip it to handle a wider range of assembly code.